feat: adds NODE_ENV, MySQL 500 error hanlder
package.json scripts now have "dev" and "prod" options.
In development, errors are passed along.  In production, we send
general errors to help give the client the right info without specifics.

This was in repsonse to the app not giving me good feedback when
the MySQL server was down.  The app was giving me 200 messages.